backend/requirements_backend.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
openpyxl
xlrd
scikit-learn
fpdf2
python-dotenv
sentence-transformers
chromadb
groq
google-generativeai
hdbscan>=0.8.33
redis>=4.0.0
supabase>=2.0.0
pgvector>=0.2.0
ollama
